// packages/core/src/generated/randart-properties.ts
export const ART_IDX = {
  BOW_SHOTS: 0,
  BOW_MIGHT: 1,
  BOW_BRAND: 2,
  BOW_SLAY: 3,
  WEAPON_HIT: 4,
  WEAPON_DAM: 5,
  NONWEAPON_HIT: 6,
  NONWEAPON_DAM: 7,
  NONWEAPON_HIT_DAM: 8,
  NONWEAPON_BRAND: 9,
  NONWEAPON_SLAY: 10,
  NONWEAPON_BLOWS: 11,
  NONWEAPON_SHOTS: 12,
  MELEE_BLESS: 13,
  MELEE_BRAND: 14,
  MELEE_SLAY: 15,
  MELEE_SINV: 16,
  MELEE_BLOWS: 17,
  MELEE_AC: 18,
  MELEE_DICE: 19,
  MELEE_WEIGHT: 20,
  MELEE_TUNN: 21,
  ALLARMOR_WEIGHT: 22,
  BOOT_AC: 23,
  BOOT_FEATHER: 24,
  BOOT_STEALTH: 25,
  BOOT_TRAP_IMM: 26,
  BOOT_SPEED: 27,
  BOOT_MOVES: 28,
  GLOVE_AC: 29,
  GLOVE_FA: 30,
  GLOVE_DEX: 31,
  GLOVE_HIT_DAM: 32,
  HELM_AC: 33,
  HELM_RBLIND: 34,
  HELM_ESP: 35,
  HELM_SINV: 36,
  HELM_WIS: 37,
  HELM_INT: 38,
  SHIELD_AC: 39,
  SHIELD_LRES: 40,
  CLOAK_AC: 41,
  CLOAK_STEALTH: 42,
  ARMOR_AC: 43,
  ARMOR_STEALTH: 44,
  ARMOR_HLIFE: 45,
  ARMOR_CON: 46,
  ARMOR_LRES: 47,
  ARMOR_ALLRES: 48,
  ARMOR_HRES: 49,
  GEN_STAT: 50,
  GEN_SUST: 51,
  GEN_STEALTH: 52,
  GEN_SEARCH: 53,
  GEN_INFRA: 54,
  GEN_SPEED: 55,
  GEN_IMMUNE: 56,
  GEN_FA: 57,
  GEN_HLIFE: 58,
  GEN_FEATHER: 59,
  GEN_LIGHT: 60,
  GEN_SINV: 61,
  GEN_ESP: 62,
  GEN_SDIG: 63,
  GEN_REGEN: 64,
  GEN_LRES: 65,
  GEN_RPOIS: 66,
  GEN_RFEAR: 67,
  GEN_RLIGHT: 68,
  GEN_RDARK: 69,
  GEN_RBLIND: 70,
  GEN_RCONF: 71,
  GEN_RSOUND: 72,
  GEN_RSHARD: 73,
  GEN_RNEXUS: 74,
  GEN_RNETHER: 75,
  GEN_RCHAOS: 76,
  GEN_RDISEN: 77,
  GEN_AC: 78,
  GEN_TUNN: 79,
  GEN_ACTIV: 80,
  GEN_PSTUN: 81,
  GEN_DAM_RED: 82,
  GEN_MOVES: 83,
  GEN_TRAP_IMM: 84,
  WEAPON_AGGR: 85,
  NONWEAPON_AGGR: 86,
  MELEE_DICE_SUPER: 87,
  BOW_SHOTS_SUPER: 88,
  BOW_MIGHT_SUPER: 89,
  GEN_SPEED_SUPER: 90,
  MELEE_BLOWS_SUPER: 91,
  MELEE_AC_SUPER: 92,
  GEN_AC_SUPER: 93,
  TOTAL: 94,
} as const;